+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/block/RedstoneInterfaceBasicBlock.kt
@@ -21,13 +21,9 @@ class RedstoneInterfaceBasicBlock : EntityBlock, Block(
 	override fun neighborChanged(state: BlockState, level: Level, pos: BlockPos, neighborBlock: Block, neighborPos: BlockPos, movedByPiston: Boolean) {
 		super.neighborChanged(state, level, pos, neighborBlock, neighborPos, movedByPiston)
 
-		val blockEntity = level.getBlockEntity(pos) as? RedstoneInterfaceBasicBlockEntity ?: return
-		val linkedPos = blockEntity.getLinkedPos() ?: return
-
-		if (level.isLoaded(linkedPos)) {
-			val linkedState = level.getBlockState(linkedPos)
-			linkedState.handleNeighborChanged(level, linkedPos, this, pos, false)
-			level.updateNeighborsAt(linkedPos, linkedState.block)
+		val blockEntity = level.getBlockEntity(pos)
+		if (blockEntity is RedstoneInterfaceBasicBlockEntity) {
+			blockEntity.updateTarget()
 		}
 	}